requirements.txt
accelerate==1.7.0
-e git+https://github.com/basedlsg/Binah.git@6b40a2ffe448f76c3812452c9604f1b3746ab21f#egg=ai_recruitment_workflow&subdirectory=LinkedInAutomaton
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
alembic==1.16.1
altair==5.5.0
amazon-braket-default-simulator==1.26.7
amazon-braket-schemas==1.24.2
amazon-braket-sdk==1.94.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.7.8
antlr4-python3-runtime==4.9.2
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.10.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
asyncio-mqtt==0.16.2
asyncio-throttle==1.0.2
asyncpg==0.30.0
attrs==25.3.0
autograd==1.8.0
autoray==0.7.1
awscli==1.40.45
babel==2.17.0
backoff==2.2.1
backports.entry-points-selectable==1.3.0
backrefs==5.8
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bitsandbytes==0.42.0
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blis==1.2.1
boltons==25.0.0
boto3==1.38.46
botocore==1.38.46
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chromadb==1.0.13
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.1
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
coverage==7.9.1
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
datasets==2.19.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diastatic-malt==2.15.2
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.19
durationpy==0.10
ecdsa==0.19.1
elevenlabs==2.3.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.104.1
fastjsonschema==2.21.1
filelock==3.18.0
flake8==7.3.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.6.2
fsspec==2024.3.1
gast==0.6.0
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-run==0.10.18
google-cloud-storage==3.1.0
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-generativeai==0.3.2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.4
humanfriendly==10.0
identify==2.6.12
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.9.0
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kombu==5.5.3
kubernetes==33.1.0
langcodes==3.5.0
langsmith==0.4.1
language_data==1.3.0
llama_api_client==0.1.1
lxml==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.3
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.12
mypy==1.15.0
mypy_extensions==1.1.0
narwhals==1.41.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.90.0
openpulse==1.0.1
openqasm3==1.0.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
oqpy==0.3.7
orjson==3.10.18
overrides==7.7.0
packaging==23.2
paginate==0.5.7
paho-mqtt==2.1.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
peft==0.15.2
PennyLane==0.41.1
PennyLane_Lightning==0.41.1
pexpect==4.9.0
pillow==10.4.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
playwright==1.52.0
plotly==5.17.0
pluggy==1.6.0
posthog==5.4.0
pre_commit==4.2.0
preshed==3.0.9
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.1
pymdown-extensions==10.16
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==26.4.0
-e git+https://github.com/basedlsg/Binah.git@6b40a2ffe448f76c3812452c9604f1b3746ab21f#egg=quest_dev_copilot&subdirectory=UNREAL-META
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.25.1
rsa==4.7.2
ruff==0.11.10
rustworkx==0.16.0
s3transfer==0.13.0
safetensors==0.5.3
schedule==1.2.0
scikit-learn==1.6.1
scipy==1.15.3
scipy-openblas32==0.3.30.0.0
seaborn==0.12.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
streamlit==1.28.1
structlog==25.4.0
sympy==1.14.0
tenacity==8.5.0
termcolor==3.1.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.0
tornado==6.5.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.52.4
typer==0.15.4
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
validators==0.35.0
vine==5.1.0
virtualenv==20.31.2
wandb==0.20.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
numpy>=1.21
matplotlib>=3.0
scipy>=1.11
